Apply further feedback
