+++ framework/src/testing/fixtures/index.ts
@@ -15,3 +15,4 @@
 
 export * from './genesis_block';
 export * from './accounts';
+export * from './config';
